udpate selenium backport #573
