+++ terasoluna-tourreservation-selenium/src/test/java/org/terasoluna/tourreservation/tourreserve/selenium/CustomerRegisterTest.java
@@ -17,12 +17,9 @@ package org.terasoluna.tourreservation.tourreserve.selenium;
 
 import static org.junit.Assert.assertEquals;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.Select;
 import org.springframework.test.context.ContextConfiguration;
@@ -35,14 +32,7 @@ import org.terasoluna.tourreservation.tourreserve.common.constants.MessageKeys;
         "classpath:META-INF/spring/seleniumContext.xml" })
 public class CustomerRegisterTest extends FunctionTestSupport {
 
-    WebDriver driver;
-
     public CustomerRegisterTest() {
-    }
-
-    @Before
-    public void setUp() {
-        driver = createWebDriver();
     }
 
     @Test
@@ -123,8 +113,4 @@ public class CustomerRegisterTest extends FunctionTestSupport {
                 "tokyo-toyosu", "********");
     }
 
-    @After
-    public void tearDown() {
-        driver.quit();
-    }
 }
