+++ terasoluna-tourreservation-selenium/src/test/java/org/terasoluna/tourreservation/tourreserve/selenium/CustomerRegisterValidateTest.java
@@ -21,12 +21,9 @@ import static org.junit.Assert.assertTrue;
 import java.util.Arrays;
 import java.util.List;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.Select;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -38,14 +35,7 @@ import org.terasoluna.tourreservation.tourreserve.common.constants.MessageKeys;
         "classpath:META-INF/spring/seleniumContext.xml" })
 public class CustomerRegisterValidateTest extends FunctionTestSupport {
 
-    WebDriver driver;
-
     public CustomerRegisterValidateTest() {
-    }
-
-    @Before
-    public void setUp() {
-        driver = createWebDriver();
     }
 
     @Test
@@ -634,8 +624,4 @@ public class CustomerRegisterValidateTest extends FunctionTestSupport {
         return true;
     }
 
-    @After
-    public void afterClass() {
-        driver.quit();
-    }
 }
