+++ terasoluna-tourreservation-selenium/src/test/java/org/terasoluna/tourreservation/tourreserve/selenium/SystemErrorTest.java
@@ -17,12 +17,9 @@ package org.terasoluna.tourreservation.tourreserve.selenium;
 
 import static org.junit.Assert.assertEquals;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.terasoluna.tourreservation.tourreserve.common.FunctionTestSupport;
@@ -33,14 +30,7 @@ import org.terasoluna.tourreservation.tourreserve.common.constants.MessageKeys;
         "classpath:META-INF/spring/seleniumContext.xml" })
 public class SystemErrorTest extends FunctionTestSupport {
 
-    WebDriver driver;
-
     public SystemErrorTest() {
-    }
-
-    @Before
-    public void setUp() {
-        driver = createWebDriver();
     }
 
     @Test
@@ -67,8 +57,4 @@ public class SystemErrorTest extends FunctionTestSupport {
                 .cssSelector("li")).getText());
     }
 
-    @After
-    public void tearDown() {
-        driver.quit();
-    }
 }
