+++ terasoluna-tourreservation-selenium/src/test/java/org/terasoluna/tourreservation/tourreserve/selenium/TokenCheckErrorTest.java
@@ -18,13 +18,10 @@ package org.terasoluna.tourreservation.tourreserve.selenium;
 import static org.junit.Assert.assertEquals;
 
 import org.joda.time.DateTime;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.Select;
 import org.springframework.test.context.ContextConfiguration;
@@ -37,14 +34,7 @@ import org.terasoluna.tourreservation.tourreserve.common.constants.MessageKeys;
         "classpath:META-INF/spring/seleniumContext.xml" })
 public class TokenCheckErrorTest extends FunctionTestSupport {
 
-    WebDriver driver;
-
     public TokenCheckErrorTest() {
-    }
-
-    @Before
-    public void setUp() {
-        driver = createWebDriver();
     }
 
     @Test
@@ -367,8 +357,4 @@ public class TokenCheckErrorTest extends FunctionTestSupport {
 
     }
 
-    @After
-    public void tearDown() {
-        driver.quit();
-    }
 }
