+++ terasoluna-tourreservation-selenium/src/test/java/org/terasoluna/tourreservation/tourreserve/selenium/TourSearchRegistValidateTest.java
@@ -20,12 +20,9 @@ import static org.junit.Assert.assertEquals;
 import javax.inject.Inject;
 
 import org.joda.time.DateTime;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.Select;
 import org.springframework.context.MessageSource;
@@ -41,14 +38,7 @@ public class TourSearchRegistValidateTest extends FunctionTestSupport {
     @Inject
     MessageSource messageSource;
 
-    WebDriver driver;
-
     public TourSearchRegistValidateTest() {
-    }
-
-    @Before
-    public void setUp() {
-        driver = createWebDriver();
     }
 
     @Test
@@ -107,8 +97,4 @@ public class TourSearchRegistValidateTest extends FunctionTestSupport {
                                         .getText());
     }
 
-    @After
-    public void tearDown() {
-        driver.quit();
-    }
 }
